Implementação de templates e rotina para envio das notificações.

+++ src/main/java/com/clau/eventntfy/service/NotificationService.java
@@ -1,17 +1,25 @@
 package com.clau.eventntfy.service;
 
 import com.clau.eventntfy.dto.request.NotificationRequestDTO;
+import com.clau.eventntfy.enums.NotificationStatus;
+import com.clau.eventntfy.enums.TypeNotification;
 import com.clau.eventntfy.mapper.NotificationMapper;
 import com.clau.eventntfy.model.Notification;
 import com.clau.eventntfy.model.User;
 import com.clau.eventntfy.repository.NotificationRepository;
 import com.clau.eventntfy.repository.UserRepository;
+import io.micrometer.common.util.StringUtils;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 @Service
@@ -22,14 +30,23 @@ public class NotificationService {
   private final UserRepository userRepository;
   private final NotificationMapper mapper;
   private final EmailService emailService;
+  private final SmsService smsService;
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(NotificationService.class);
+
+  private final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm");
 
   public void saveNotification(NotificationRequestDTO requestDTO) {
     if (requestDTO == null) {
       throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Notificação não informada.");
     }
 
-    if (requestDTO.getScheduledTime().isBefore(LocalDateTime.now())) {
+    if (requestDTO.getScheduledTime().isBefore(ZonedDateTime.now(ZoneId.of("America/Sao_Paulo")).toLocalDateTime())) {
       throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Data agendada não pode ser anterior à data atual.");
+    }
+
+    if (StringUtils.isBlank(requestDTO.getSubject()) && TypeNotification.EMAIL.equals(requestDTO.getType())) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Assunto do email não informado.");
     }
 
     User sender = userRepository.findById(requestDTO.getUserId())
@@ -46,9 +63,13 @@ public class NotificationService {
 
     Notification notificationEntity = repository.save(notification);
 
-    emailService.sendEmail(notificationEntity);
-
-    // TODO: Implementar envio de e-mail informando a criação da notificação.
+    if (TypeNotification.EMAIL.equals(notificationEntity.getType())) {
+      LOGGER.info("Enviando email de CRIAÇÃO para os destinatários...");
+      emailService.sendEmail(notificationEntity);
+    } else if (TypeNotification.SMS.equals(notificationEntity.getType())) {
+      LOGGER.info("Enviando SMS de CRIAÇÃO para os destinatários...");
+      smsService.sendSms(notification);
+    }
   }
 
   public void deleteNotification(Long id) {
@@ -59,9 +80,51 @@ public class NotificationService {
     Notification notification = repository.findById(id)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Notificação não encontrada."));
 
+    if(notification.getStatus() == NotificationStatus.SENT) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Não é possível cancelar uma notificação já enviada.");
+    }
+
+    if(notification.getStatus() == NotificationStatus.FAILED) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Não é possível cancelar uma notificação já cancelada.");
+    }
+
+    notification.setStatus(NotificationStatus.FAILED);
+    notification.setMessage(("O Evento foi cancelado %s \n " +
+            "%s").formatted(ZonedDateTime.now(ZoneId.of("America/Sao_Paulo")).format(dateTimeFormatter), notification.getMessage()));
+
+    if (TypeNotification.EMAIL.equals(notification.getType())) {
+      LOGGER.info("Enviando email de cancelamento para os destinatários...");
+      emailService.sendEmail(notification);
+    } else if (TypeNotification.SMS.equals(notification.getType())) {
+      LOGGER.info("Enviando SMS de cancelamento para os destinatários...");
+      smsService.sendSms(notification);
+    }
+
     notification.getRecipients().clear();
     repository.delete(notification);
-
-    // TODO: Implementar envio de e-mail informando a exclusão da notificação.
   }
+
+  public void sendNotifications() {
+
+    ZonedDateTime saoPauloTime = ZonedDateTime.now(ZoneId.of("America/Sao_Paulo"));
+    LocalDateTime now = saoPauloTime.toLocalDateTime();
+
+    List<Notification> notifications = repository.findAllPendingNotifications(now);
+
+    notifications.forEach(notification -> {
+      notification.setSent(true);
+      notification.setStatus(NotificationStatus.SENT);
+
+      if (TypeNotification.EMAIL.equals(notification.getType())) {
+        LOGGER.info("Enviando email de agendamento para os destinatários...");
+        emailService.sendEmail(notification);
+      } else if (TypeNotification.SMS.equals(notification.getType())) {
+        LOGGER.info("Enviando SMS de agendamento para os destinatários...");
+        smsService.sendSms(notification);
+      }
+
+      repository.save(notification);
+    });
+  }
+
 }
